elimu_backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
pillow==11.3.0
pymongo==4.15.2
djongo==1.3.6
sqlparse==0.2.4

# API Documentation
drf-yasg==1.21.7

# Environment variables
python-decouple==3.8
django-environ==0.11.2

# Additional utilities
django-filter==24.2
dnspython==2.8.0

# Media Storage - Cloudinary
cloudinary==1.40.0
django-cloudinary-storage==0.3.0

# PDF Generation
weasyprint==62.3
reportlab==4.2.5

# Background Tasks (Optional)
celery==5.4.0
redis==5.0.3

# Utilities
python-slugify==8.0.1

